+++ src/types/payments.ts
@@ -1,18 +1,28 @@
 import { MiscTypes, SubmissionEventTypes } from '@oneblink/types'
 import { FormSubmissionResult } from './submissions'
 
+export type PaymentReceiptItem = {
+  /** The label to represent the value */
+  label: string
+  /** The value to display */
+  value: string
+  /** A CSS class to add to the HTML element container */
+  className?: string
+  /** A CSS class to add to the HTML element containing the value */
+  valueClassName?: string
+  /** A [material icon](https://fonts.google.com/icons) */
+  icon?: string
+  /** Display a button on the receipt item to copy the value to the clipboard */
+  allowCopyToClipboard: boolean
+}
+
 export type HandlePaymentResult = {
+  receiptItems: PaymentReceiptItem[]
   transaction: {
     /** `true` if the transaction was successful */
     isSuccess: boolean
     /** The error message to display if `isSuccess` is `false` */
     errorMessage: string | MiscTypes.NoU
-    /** The id the transaction */
-    id: string | MiscTypes.NoU
-    /** A mask of the credit card used e.g. _1234....7890_ */
-    creditCardMask: string | MiscTypes.NoU
-    /** The total amount charged */
-    amount: number | MiscTypes.NoU
   }
   submissionResult: FormSubmissionResult
 }